no time to finish compile, fixes to cxx warnings
git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@1050 a592a061-630c-0410-9148-cb99ea01b6c8

+++ src/Painter.C
@@ -269,7 +269,7 @@ PainterBase & Painter::text(int x, int y, char c, LyXFont const & f)
 }
 
 
-PainterBase & Painter::text(int x, int y, char const * s, int ls,
+PainterBase & Painter::text(int x, int y, char const * s, size_t ls,
 			LyXFont const & f)
 {
 	if (lyxrc.font_norm_type == LyXRC::ISO_10646_1) {
